Delete SnakeGame.jack
I know I have been saying  this a lot but I feel this folder should be deleted. I've seen students use this and claim as their own which is cheating and this comes from Nand2tetris which is used in many classrooms around the country so a lot of people can just look online and just find it, make some small changes and then call it as their own. I feel that this really should be deleted. People's work like this shouldn't be taken for granted and thus should be deleted so that students can actually do their own work and show their true potential.